miscellaneous improvements to HLTriggerOffline/Scouting plugins
